requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
django-environ==0.11.2
django-filter==24.2
django-password-validators==1.7.3
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
greenlet==3.0.3
idna==3.7
iniconfig==2.0.0
Markdown==3.6
packaging==24.1
pillow==10.3.0
playwright==1.44.0
pluggy==1.5.0
pyee==11.1.0
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-base-url==2.1.0
pytest-django==4.8.0
pytest-playwright==0.5.0
python-slugify==8.0.4
requests==2.32.3
sqlparse==0.5.0
text-unidecode==1.3
typing_extensions==4.12.2
urllib3==2.2.1
